addon_callable_length
Add a new addon pair, scriptlength.cpp/h, which register a value class addon_callable_length. This class acts like a uint in as many ways as possible but also returns itself when called. When this class is registered, string and array objects ignore the value of AS_USE_ACCESSORS and instead return an addon instance for their get_length properties, meaning that arr.length and arr.length() are both valid again, as they were in earlier releases of AngelScript. Using length as a setter (as an alternative to resize), however, is impossible.